+++ app/src/main/java/com/hampshirewolves/aeroatlas/ui/mainactivity/MainActivityViewModel.java
@@ -7,9 +7,7 @@ import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 
 import com.hampshirewolves.aeroatlas.model.City;
-import com.hampshirewolves.aeroatlas.model.CityAttractions;
-import com.hampshirewolves.aeroatlas.model.CityAttractionsRepository;
-import com.hampshirewolves.aeroatlas.model.CityRepository;
+import com.hampshirewolves.aeroatlas.repository.CityRepository;
 
 import java.util.List;
