+++ app/src/main/java/com/hampshirewolves/aeroatlas/ui/mainactivity/RecyclerViewInterface.java
@@ -1,4 +1,4 @@
-package com.hampshirewolves.aeroatlas.homepage;
+package com.hampshirewolves.aeroatlas.ui.mainactivity;
 
 public interface RecyclerViewInterface {
     void onItemClick(int position);
