+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@ xmlns:tools="http://schemas.android.com/tools"
 android:id="@+id/main"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".LoginActivity">
+tools:context=".ui.LoginActivity">
 
 <androidx.cardview.widget.CardView
     android:id="@+id/card_view"
