+++ app/src/main/res/layout/activity_signup.xml
@@ -4,7 +4,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignupActivity">
+    tools:context=".ui.SignupActivity">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/card_view"
