+++ app/src/main/res/layout/fragment_homepage.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="12dp"
-    tools:context=".homepage.HomepageFragment">
+    tools:context=".ui.fragments.homepage.HomepageFragment">
 
     <SearchView
         android:id="@+id/homepage_city_search_bar"
