+++ app/src/main/res/layout/fragment_user_page.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".profilepage.UserPageFragment">
+    tools:context=".ui.fragments.UserPageFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
